requirements.txt
pytest
psycopg2-binary
fastapi
uvicorn